requirements.txt
numpy
tensorflow-gpu==1.4.0
nltk==3.2.3
